remove abstract solution loader check for result version
