+++ tests/testcases/run_test.py
@@ -521,6 +521,4 @@ if __name__ == "__main__":
 
     config.solver.keep_files = False
     folder_path = os.path.dirname(__file__)
-
-    test_1a(config, folder_path)
-    a=1
+    test_8a(config, folder_path)
